Moved trait name validation function to utils.py
